Fix integer overflow in isPrime loop condition and add Integer.MAX_VALUE test

+++ PrimeNumberChecker.java
@@ -26,7 +26,7 @@ public class PrimeNumberChecker {
         if (number % 2 == 0 || number % 3 == 0) {
             return false;
         }
-        for (int i = 5; i * i <= number; i += 6) {
+        for (int i = 5; (long) i * i <= number; i += 6) {
             if (number % i == 0 || number % (i + 2) == 0) {
                 return false;
             }
